Add data conversion

+++ lib/nnc/cmd/util/ccv_nnc_util_cpu_ref.c
@@ -358,7 +358,7 @@ static int _ccv_nnc_datatype_conversion(const ccv_nnc_cmd_t cmd, const ccv_nnc_h
 		} else if (a->info.datatype == CCV_32F && b->info.datatype == CCV_16F) {
 			assert(!CCV_IS_TENSOR_VIEW(a));
 			assert(!CCV_IS_TENSOR_VIEW(b));
-			const int tensor_count = ccv_nnc_tensor_count(a->info);
+			const size_t tensor_count = ccv_nnc_tensor_count(a->info);
 			assert(tensor_count == ccv_nnc_tensor_count(b->info));
 			ccv_float_to_half_precision(a->data.f32, (uint16_t*)b->data.f16, tensor_count);
 		} else if (a->info.datatype == CCV_16F && b->info.datatype == CCV_32F) {
